requirements.txt
tensorflow
numpy
matplotlib
opencv-python
scikit-learn
flask
flask-cors
pillow